requirements.txt
snowflake-connector-python
elasticsearch
PyYAML
python-json-logger
python-dotenv
python-dateutil
